# periph/hwrng: Introduce generic hwrng_read(), allow for leaner HWRNG implementations