--- a/cpu/esp32/periph/hwrng.c
+++ b/cpu/esp32/periph/hwrng.c
@@ -23,33 +23,9 @@
 #include "periph_conf.h"
 #include "periph/hwrng.h"
 
-static const uint32_t* RNG_DATA_REG = (uint32_t*)0x3ff75144;
+#define RNG_DATA_REG (*(volatile uint32_t *)0x3ff75144)
 
-void hwrng_init(void)
+uint32_t hwrng_uint32(void)
 {
-    /* no need for initialization */
-}
-#include "rom/ets_sys.h"
-void hwrng_read(void *buf, unsigned int num)
-{
-    unsigned int count = 0;
-    uint8_t *b = (uint8_t *)buf;
-
-    while (count < num) {
-        /* read next 4 bytes of random data */
-        uint32_t tmp = *RNG_DATA_REG;
-
-        /* copy data into result vector */
-        for (int i = 0; i < 4 && count < num; i++) {
-            b[count++] = (uint8_t)tmp;
-            tmp = tmp >> 8;
-        }
-    }
-}
-
-uint32_t hwrand (void)
-{
-    uint32_t _tmp;
-    hwrng_read(&_tmp, sizeof(uint32_t));
-    return _tmp;
+    return RNG_DATA_REG;
 }