--- a/cpu/esp8266/periph/hwrng.c
+++ b/cpu/esp8266/periph/hwrng.c
@@ -25,28 +25,6 @@
 
 #include "esp/wdev_regs.h"
 
-void hwrng_init(void)
-{
-    /* no need for initialization */
-}
-
-void hwrng_read(void *buf, unsigned int num)
-{
-    unsigned int count = 0;
-    uint8_t *b = (uint8_t *)buf;
-
-    while (count < num) {
-        /* read next 4 bytes of random data */
-        uint32_t tmp = WDEV.HWRNG;
-
-        /* copy data into result vector */
-        for (int i = 0; i < 4 && count < num; i++) {
-            b[count++] = (uint8_t)tmp;
-            tmp = tmp >> 8;
-        }
-    }
-}
-
 uint32_t hwrng_uint32(void)
 {
     return WDEV.HWRNG;