--- a/cpu/mips_pic32_common/include/periph_cpu_common.h
+++ b/cpu/mips_pic32_common/include/periph_cpu_common.h
@@ -39,6 +39,12 @@ extern "C" {
 #define CPUID_LEN           (4U)
 
 /**
+ * @name    HWRNG configuration
+ * @{
+ */
+#define HWRNG_HAS_INIT   (1)
+
+/**
  * @brief   Override GPIO pin selection macro
  */
 #define GPIO_PIN(x,y)       ((x << 4) | (y & 0xf))