--- a/cpu/sam0_common/periph/hwrng.c
+++ b/cpu/sam0_common/periph/hwrng.c
@@ -32,25 +32,8 @@ void hwrng_init(void)
     TRNG->CTRLA.bit.ENABLE = 1;
 }
 
-uint32_t hwrand(void)
+uint32_t hwrng_uint32(void)
 {
     while (!TRNG->INTFLAG.bit.DATARDY) {}
     return TRNG->DATA.reg;
-}
-
-void hwrng_read(void *buf, unsigned int num)
-{
-    unsigned int count = 0;
-    uint8_t *b = (uint8_t *)buf;
-
-    while (count < num) {
-        /* read next 4 bytes of random data */
-        uint32_t tmp = hwrand();
-
-        /* copy data into result vector */
-        for (int i = 0; i < 4 && count < num; i++) {
-            b[count++] = (uint8_t)tmp;
-            tmp = tmp >> 8;
-        }
-    }
 }