--- a/drivers/periph_common/hwrng.c
+++ b/drivers/periph_common/hwrng.c
@@ -17,6 +17,7 @@
  * @}
  */
 
+#include <stddef.h>
 #include "periph_cpu.h"
 #include "periph/hwrng.h"