### Contribution description
Most HWRNGs are pretty simple. They provide a register that contains random data. Some HWRNG implementations already expose this single register by providing a `hwrand(void)` function that returns a 32 bit value.

The `hwrng_read()` API function expects a buffer, so currently many implementations invent (or copy & paste) their way to fill a buffer of arbitrary size with 32 bit values.

This PR does the following things:
 - add the `hwrand(void)` function to the hwrng API. Not only is it often more convenient, it also allows for a generic `hwrng_read()` implementation.
 - implement `hwrand()` for all drivers that didn't have it yet
 - implement a generic `hwrng_read()` that uses `hwrand()` and is a bit more efficient than the current implementation
 - Introduce `HWRNG_HAS_INIT` & `HWRNG_HAS_POWERONOFF` analogous to the watchdog API so we don't have to call empty functions if they are not needed.

### Testing procedure

All platforms should still build & run `tests/rng` and `tests/periph_hwrng` as they did before.

On same54 a small speedup can be observed using the `speed` test in `tests/rng` (set `source 2` to select the HWRNG):

before: `Collected 11009147 samples in 10.000008 seconds (4300 KiB/s).`
after: `Collected 12244868 samples in 10.000009 seconds (4783 KiB/s).`


### Issues/PRs references

<!--
Examples: Fixes #1234. See also #5678. Depends on PR #9876.

Please use keywords (e.g., fixes, resolve) with the links to the issues you
resolved, this way they will be automatically closed when your pull request
is merged. See https://help.github.com/articles/closing-issues-using-keywords/.
-->