@maribu 2019-10-22
@haukepetersen and @mehlis were the authors of the API, so maybe they want to join the discussion.

I personally do not like the idea of exposing hardware random number generators directly to the application programmer. I'm pretty sure that the use case is not "talk to a HWRNG" but "get true random". And "talk to a HWRNG" is one (very nice) way to implement "get true random", but not the only way.

With that in mind, I would personally prefer a user facing API like this:

``` C
typedef enum {
    GETENTROPY_MODE_BLOCK, /**< Block until the required amount of entropy was obtained */
    GETENTROPY_MODE_DO_NOT_BLOCK, /**< Get as much entropy as is available */
    /**
     * @brief Try to give the required amount of entropy with first level entroy for some
     *        time. If this fails, return pseudo-random values to avoid blocking
     */
    GETENTROPY_MODE_UNSECURE_FALLBACK,
} getentropy_mode_t;

/**
 * @brief Retrieve entropy gathered and mixed from all available entropy sources
 *
 * @param dest Store the entropy in this buffer
 * @param size_t The number of entropy bytes to retrieve
 * @param mode What to do if the required amount of entropy is not available
 *
 * @retval -1 Failed to get entropy
 * @return The number of entropy bytes actually retrieved, which can be less than the
 *         requested number depending on the value of @p mode
 *
 * This function is safe to call concurrently.
 *
 * @details Internally a mutex must be used to ensure that the same entropy is only
 *          delivered to a single context. Calls with @p mode set to
 *          @ref ENTROPY_MODE_DO_NOT_BLOCK will try to lock that mutex and return no
 *          entropy, if the mutex is already locked.
 */
ssize_t getentropy(void *dest, size_t length, getentroopy_mode_t mode);

/**
 * @brief Initialize the entropy backend
 *
 * @pre   This function is only called once. If the module `auto_init` is used, it will
 *        be called from there
 */
void entropy_init(void);
```

(A set of convenience functions to get a truly random `uint32_t` value etc. could be provided, if there is interest in having them. But I personally can only see two use cases for true randomness: Generating key or IVs for crypto stuff, or seeding a PRNG. In both cases those convenience functions are not really useful.)

This API could than be implemented in the most ROM and RAM efficient way by only having a single HWRNG as entropy source. As this device will never block indefinitely, `getentropy()` could just ignore the `mode` parameter and always return the amount of requested entropy. (And this should definitely a choice the user can pick from, when the device has a periph HWRNG.)

If however the user decides to not trust the HWRNG to be fully random or does not have any HWRNG , a more advanced implementation of that API could be used that mixes many sources for entropy in. That way the damage of a faulty (or backdoored) entropy source would be limited. I see two classes of providers for this entropy: 

1. On-demand sources like HWRNGs or transceivers that can read the RSSI of background noise. Those can be called from `getentropy()`.
2. Occasional entropy sources. E.g. the timestamp of any external interrupt will have some entropy, as even for periodic interrupts some jitter will occur. (I don't this that timer interrupts of timers fed by the same clock as the CPU will add entropy, but every other interrupt should do, I guess.) Other sources might by transceivers to only provide the RSSI of received frames. Those sources will have to call an internal API to feed their entropy into the mix whenever an entropy generating event takes places.

The advantages of the API I proposed above that I see are:

1. Multiple sources for entropy can (***optionally!***) be used and mixed in to increase the "robustness" against [hardware failures](https://lore.kernel.org/patchwork/patch/1115413/) and [backdoored entropy sources](https://en.wikipedia.org/wiki/Global_surveillance_disclosures_(2013%E2%80%93present))
2. Not only "on-demand" entropy sources like hardware random number generators could be accessed in the same API. E.g. if we have a RIOT node without HWRNGs, ADCs (that can be used to sample noise), transceivers (that allows harvesting RSSI background noise) but still needs to generate crypto keys - lets say a KNX connected smart home device: The new API could still cater this use case with occasional entropy sources like external interrupts.