@haukepetersen 2019-10-22
@maribu this periph interface is not really meant to be a generic user facing API that is able to cover multiple different types of entropy sources (rssi, ...), but a slim API that only wraps the on-cpu hardware random number peripherals on CPUs (if present). The 'multiplexing' of different entropy sources etc is/should be actually part of the higher level `random` API.