@maribu 2019-10-22
> @maribu this periph interface is not really meant to be a generic user facing API

OK, in that case I'm fine. Sorry for making the noise.

> The 'multiplexing' of different entropy sources etc is/should be actually part of the higher level `random` API.

That API exists, but "only" provides PRNGs. To me it makes sense to add a second higher level API for true random (as needed for cryptographic keys). Keeping those two strictly apart makes sense, as they have totally different requirements. E.g. I could use a PRNG for visual or sound effects where actual randomness doesn't matter, but speed a lot.