@miri64 2019-10-22
> That API exists, but "only" provides PRNGs. To me it makes sense to add a second higher level API for true random (as needed for cryptographic keys). Keeping those two strictly apart makes sense, as they have totally different requirements. E.g. I could use a PRNG for visual or sound effects where actual randomness doesn't matter, but speed a lot.

![](https://media.giphy.com/media/hc10gBL10d3Ko/giphy.gif)

#scnr the random API got so often (re-)discussed in the past ^^