@kaspar030 2019-10-22
> #scnr the random API got so often (re-)discussed in the past ^^

I thought the same. :wink: But I don't find the discussion with the actual API proposals that got some consensus but never got realized.