@benpicco 2019-10-22
@haukepetersen my main motivation was to streamline the implementation because it irked me there was so much duplicated code for all the hwrng implementations.
Then I discovered `tests/rng` has a benchmark mode…

Adding `hwrng_uint16()` and `hwrng_uint8()` would be worthwhile too. Because I disagree with

> But I personally can only see two use cases for true randomness: Generating key or IVs for crypto stuff, or seeding a PRNG. In both cases those convenience functions are not really useful.)

A PRNG just adds unnecessary code size, if a HWRNG exists why not just use that for all random needs? That's why I wanted it to have a convenient API and why I wanted to avoid having the overhead of enabling / disabling the peripheral for each `hwrng_uint32()` call.