@maribu 2019-10-22
> I thought the same. wink But I don't find the discussion with the actual API proposals that got some consensus but never got realized.

Please share if you can find the discussion. I'd like to finally create a PR for RIOT that is actually related to my field of research :laughing: